+++ retz-admin/src/main/java/io/github/retz/admin/CommandEnableUser.java
@@ -43,7 +43,8 @@ public class CommandEnableUser implements SubCommand {
 
     @Override
     public int handle(FileConfiguration fileConfig) throws Throwable {
-        try(JmxClient jmxClient = new JmxClient("localhost", 9999)) {
+        int port = fileConfig.getJmxPort();
+        try(JmxClient jmxClient = new JmxClient("localhost", port)) {
             Object o = jmxClient.invokeOperation(new ObjectName("io.github.retz.scheduler:type=AdminConsole"), "enableUser", id, true);
             boolean result = (Boolean)o;
             LOG.info("User enabled: {}", result);