+++ retz-admin/src/main/java/io/github/retz/admin/CommandGetUser.java
@@ -44,7 +44,8 @@ public class CommandGetUser implements SubCommand {
 
     @Override
     public int handle(FileConfiguration fileConfig) throws Throwable  {
-        try (AdminConsoleClient client = new AdminConsoleClient(new JmxClient("localhost", 9999))) {
+        int port = fileConfig.getJmxPort();
+        try (AdminConsoleClient client = new AdminConsoleClient(new JmxClient("localhost", port))) {
             User user = client.getUserAsObject(id);
             LOG.info("User id={}, enabled={}", user.keyId(), user.enabled());
             return 0;