+++ retz-admin/src/main/java/io/github/retz/admin/CommandListUser.java
@@ -41,7 +41,8 @@ public class CommandListUser implements SubCommand {
 
     @Override
     public int handle(FileConfiguration fileConfig) throws Throwable {
-        try(AdminConsoleClient client = new AdminConsoleClient(new JmxClient("localhost", 9999))) {
+        int port = fileConfig.getJmxPort();
+        try(AdminConsoleClient client = new AdminConsoleClient(new JmxClient("localhost", port))) {
             List<String> users = client.listUser();
             for(String id: users) {
                 LOG.info(id);